trial 94: cleanup pass — extract 3 shared utils, delete 805 lines of dead/duplicated code
Net diff: +75 / -880 across 17 rule files + 3 new util files.

## Extracted shared utils

New: `is-initial-only-prop-name.ts` (25 lines) — single source of truth
for the `^initial[A-Z]` / `^default[A-Z]` / `^seed[A-Z]` / `initialValue`
prop-name heuristic. Was duplicated verbatim across no-derived-state,
no-derived-state-effect, and no-derived-use-state.

New: `is-namespaced-api-call.ts` (135 lines) — single source of truth
for the `editor.commands.X(...)` / `app.store.dispatch(...)` /
`editor.fonts.X(...)` namespace-method detection. Was duplicated verbatim
across no-pass-data-to-parent and no-pass-live-state-to-parent (~125
lines each).

New: `jsx-stateless-leaf.ts` (190 lines) — single source of truth for
`PURE_SVG_PRIMITIVE_TAGS` + `STATELESS_HTML_LEAF_TAGS` +
`containsStatefulDescendant` AST scan. Was duplicated verbatim across
no-array-index-key and no-array-index-as-key (~190 lines each).

## Dead code removed

- `js-combine-iterations`: removed `KEY_BOUNDED_RECEIVER_METHODS` set +
  `isKeyBoundedReceiver` helper (vestige of the reverted Object.X
  bounded-receiver skip from trial 93). Trimmed comment bloat in the
  rule body (~30 lines).

- `no-event-handler`: consolidated 3 verbose comment blocks into single
  one-line summaries; removed redundant "see comment above" trailing
  notes (~85 lines).

## Tightened over-broad patterns

- `only-export-components`:
  - `NODE_DEFINITION_BASENAME_PATTERN`: removed `Component`, `Block`,
    `Node`, `InlineNode`, `Extension`, `Plugin` from the suffix list
    (too generic — `MyComponent.tsx` shouldn't auto-skip
    only-export-components).
  - `UTILITY_BASENAME_SUFFIX_PATTERN`: removed 50+ generic
    component-type suffixes (Modal/Dialog/Drawer/Popover/Menu/Card/Tile/
    Sidebar/Toolbar/Breadcrumb/Banner/Alert/Toast/Avatar/Badge/etc.)
    that routinely ARE the single-component file the rule wants to
    protect. Kept only utility/registry/framework-specific suffixes.

## Deduped slot lists

- `jsx-no-jsx-as-prop`: removed `"modal"` dup, removed `"icon2"` typo.
- `jsx-no-new-array-as-prop`: removed `"Stages"` dup.
- `jsx-no-new-object-as-prop`: removed `"viewport"`, `"nodeTypes"`,
  `"Padding"`, `"Margin"` dups.
- `jsx-no-new-function-as-prop`: removed `"onCommit"`, `"onReset"` dups.

## Eval scripts

Deleted `react-doctor-evals/scripts/retrial-modified.mjs` (superseded by
`retrial-this-session.mjs` — same logic, different rule list).

## Test status

oxlint-plugin-react-doctor: 5,980 pass, 47 skipped (no regression).
react-doctor: 1,414 pass, 3 skipped (no regression).
