+++ packages/oxlint-plugin-react-doctor/src/plugin/rules/correctness/no-array-index-as-key.ts
@@ -5,196 +5,13 @@ import type { Rule } from "../../utils/rule.js";
 import type { RuleContext } from "../../utils/rule-context.js";
 import { isNodeOfType } from "../../utils/is-node-of-type.js";
 import type { EsTreeNodeOfType } from "../../utils/es-tree-node-of-type.js";
+import {
+  containsStatefulDescendant,
+  PURE_SVG_PRIMITIVE_TAGS,
+  STATELESS_HTML_LEAF_TAGS,
+} from "../../utils/jsx-stateless-leaf.js";
 
 const STRING_COERCION_FUNCTIONS = new Set(["String", "Number"]);
-
-// Mirror of `no-array-index-key`'s pure-SVG-primitive list — these
-// hold no DOM state, so an index-key misidentification has no
-// observable consequence.
-const PURE_SVG_PRIMITIVE_TAGS: ReadonlySet<string> = new Set([
-  "circle",
-  "ellipse",
-  "g",
-  "line",
-  "path",
-  "polygon",
-  "polyline",
-  "rect",
-  "stop",
-  "text",
-  "tspan",
-  "defs",
-  "use",
-  "mask",
-  "marker",
-  "linearGradient",
-  "radialGradient",
-  "clipPath",
-  "filter",
-  "feGaussianBlur",
-  "feOffset",
-  "feMerge",
-  "feMergeNode",
-  "feColorMatrix",
-  "feFlood",
-  "feComposite",
-  "title",
-  "desc",
-]);
-
-// Mirror of `no-array-index-key`'s stateless-leaf list. See that file
-// for the full rationale.
-const STATELESS_HTML_LEAF_TAGS: ReadonlySet<string> = new Set([
-  "div",
-  "span",
-  "p",
-  "h1",
-  "h2",
-  "h3",
-  "h4",
-  "h5",
-  "h6",
-  "header",
-  "footer",
-  "section",
-  "article",
-  "aside",
-  "main",
-  "nav",
-  "li",
-  "ul",
-  "ol",
-  "dl",
-  "dt",
-  "dd",
-  "tr",
-  "td",
-  "th",
-  "tbody",
-  "thead",
-  "tfoot",
-  "table",
-  "caption",
-  "colgroup",
-  "col",
-  "strong",
-  "em",
-  "small",
-  "b",
-  "i",
-  "u",
-  "s",
-  "mark",
-  "del",
-  "ins",
-  "sub",
-  "sup",
-  "abbr",
-  "cite",
-  "code",
-  "kbd",
-  "samp",
-  "pre",
-  "blockquote",
-  "q",
-  "br",
-  "hr",
-  "wbr",
-  "img",
-  "picture",
-  "figure",
-  "figcaption",
-  "label",
-  "legend",
-  "fieldset",
-  "address",
-  "time",
-  "data",
-  "var",
-  "ruby",
-  "rt",
-  "rp",
-  "bdi",
-  "bdo",
-]);
-
-const STATEFUL_HTML_DESCENDANT_TAGS: ReadonlySet<string> = new Set([
-  "input",
-  "textarea",
-  "select",
-  "option",
-  "optgroup",
-  "button",
-  "form",
-  "output",
-  "progress",
-  "meter",
-  "video",
-  "audio",
-  "source",
-  "track",
-  "iframe",
-  "embed",
-  "object",
-  "a",
-  "details",
-  "summary",
-  "dialog",
-  "canvas",
-]);
-
-const STATEFUL_DESCENDANT_SCAN_BUDGET = 200;
-
-const containsStatefulDescendant = (jsxElement: EsTreeNode): boolean => {
-  let budget = STATEFUL_DESCENDANT_SCAN_BUDGET;
-  const stack: Array<EsTreeNode> = [jsxElement];
-  while (stack.length > 0) {
-    if (budget <= 0) return true;
-    budget -= 1;
-    const node = stack.pop()!;
-    if (isNodeOfType(node, "JSXElement")) {
-      const opening = (node as { openingElement: EsTreeNode }).openingElement;
-      const name = (opening as { name?: EsTreeNode }).name;
-      if (name && isNodeOfType(name, "JSXIdentifier")) {
-        const tagName = name.name;
-        const firstChar = tagName.charCodeAt(0);
-        const isUppercase = firstChar >= 65 && firstChar <= 90;
-        if (isUppercase) return true;
-        if (STATEFUL_HTML_DESCENDANT_TAGS.has(tagName)) return true;
-      }
-      if (name && isNodeOfType(name, "JSXMemberExpression")) return true;
-      const children = (node as { children?: ReadonlyArray<EsTreeNode> }).children ?? [];
-      for (const child of children) stack.push(child);
-      continue;
-    }
-    if (isNodeOfType(node, "JSXFragment")) {
-      const children = (node as { children?: ReadonlyArray<EsTreeNode> }).children ?? [];
-      for (const child of children) stack.push(child);
-      continue;
-    }
-    if (isNodeOfType(node, "JSXExpressionContainer")) {
-      const expression = node.expression as EsTreeNode;
-      if (
-        isNodeOfType(expression, "CallExpression") ||
-        isNodeOfType(expression, "Identifier") ||
-        isNodeOfType(expression, "MemberExpression")
-      ) {
-        return true;
-      }
-      stack.push(expression);
-      continue;
-    }
-    if (isNodeOfType(node, "ConditionalExpression")) {
-      stack.push(node.consequent as EsTreeNode, node.alternate as EsTreeNode);
-      continue;
-    }
-    if (isNodeOfType(node, "LogicalExpression")) {
-      stack.push(node.left as EsTreeNode, node.right as EsTreeNode);
-      continue;
-    }
-  }
-  return false;
-};
 
 const extractIndexName = (node: EsTreeNode): string | null => {
   if (isNodeOfType(node, "Identifier") && INDEX_PARAMETER_NAMES.has(node.name)) return node.name;
