+++ packages/oxlint-plugin-react-doctor/src/plugin/rules/react-builtins/jsx-no-new-array-as-prop.ts
@@ -276,7 +276,6 @@ const DATA_ARRAY_PROP_SUFFIXES: ReadonlyArray<string> = [
   "Deployments",
   "Jobs",
   "Tasks",
-  "Stages",
   "Phases",
   "Milestones",
   "Cards",
