+++ packages/oxlint-plugin-react-doctor/src/plugin/rules/react-builtins/no-array-index-key.ts
@@ -3,6 +3,11 @@ import type { EsTreeNode } from "../../utils/es-tree-node.js";
 import type { EsTreeNodeOfType } from "../../utils/es-tree-node-of-type.js";
 import { hasJsxPropIgnoreCase } from "../../utils/has-jsx-prop-ignore-case.js";
 import { isNodeOfType } from "../../utils/is-node-of-type.js";
+import {
+  containsStatefulDescendant,
+  PURE_SVG_PRIMITIVE_TAGS,
+  STATELESS_HTML_LEAF_TAGS,
+} from "../../utils/jsx-stateless-leaf.js";
 import type { Rule } from "../../utils/rule.js";
 
 const MESSAGE =
@@ -252,233 +257,12 @@ const isReactCloneElement = (callExpression: EsTreeNodeOfType<"CallExpression">)
   return isNodeOfType(callee.object, "Identifier") && callee.object.name === "React";
 };
 
-// Pure-presentational SVG primitives — no DOM state, no event-bound
-// identity to corrupt on reorder. `<g>`, `<path>`, `<line>`, etc.
-// just draw pixels; React will diff the new attributes regardless of
-// reconciliation order, and there's nothing to "lose" if React maps
-// the wrong index to the wrong element. Real-world icon / chart code
-// uses index keys here as the natural choice.
-const PURE_SVG_PRIMITIVE_TAGS: ReadonlySet<string> = new Set([
-  "circle",
-  "ellipse",
-  "g",
-  "line",
-  "path",
-  "polygon",
-  "polyline",
-  "rect",
-  "stop",
-  "text",
-  "tspan",
-  "defs",
-  "use",
-  "mask",
-  "marker",
-  "linearGradient",
-  "radialGradient",
-  "clipPath",
-  "filter",
-  "feGaussianBlur",
-  "feOffset",
-  "feMerge",
-  "feMergeNode",
-  "feColorMatrix",
-  "feFlood",
-  "feComposite",
-  "title",
-  "desc",
-]);
+const isPureSvgPrimitiveJsxName = (jsxOpeningName: EsTreeNode): boolean =>
+  isNodeOfType(jsxOpeningName, "JSXIdentifier") && PURE_SVG_PRIMITIVE_TAGS.has(jsxOpeningName.name);
 
-const isPureSvgPrimitiveJsxName = (jsxOpeningName: EsTreeNode): boolean => {
-  if (!isNodeOfType(jsxOpeningName, "JSXIdentifier")) return false;
-  return PURE_SVG_PRIMITIVE_TAGS.has(jsxOpeningName.name);
-};
-
-// Plain-display HTML elements that carry no DOM-managed state
-// (selection, focus, scroll, video time, etc.) and emit no events
-// that would benefit from stable identity. Reorder hazards for these
-// only matter when their CHILDREN have state — see the descendant
-// scan below.
-const STATELESS_HTML_LEAF_TAGS: ReadonlySet<string> = new Set([
-  "div",
-  "span",
-  "p",
-  "h1",
-  "h2",
-  "h3",
-  "h4",
-  "h5",
-  "h6",
-  "header",
-  "footer",
-  "section",
-  "article",
-  "aside",
-  "main",
-  "nav",
-  "li",
-  "ul",
-  "ol",
-  "dl",
-  "dt",
-  "dd",
-  "tr",
-  "td",
-  "th",
-  "tbody",
-  "thead",
-  "tfoot",
-  "table",
-  "caption",
-  "colgroup",
-  "col",
-  "strong",
-  "em",
-  "small",
-  "b",
-  "i",
-  "u",
-  "s",
-  "mark",
-  "del",
-  "ins",
-  "sub",
-  "sup",
-  "abbr",
-  "cite",
-  "code",
-  "kbd",
-  "samp",
-  "pre",
-  "blockquote",
-  "q",
-  "br",
-  "hr",
-  "wbr",
-  "img",
-  "picture",
-  "figure",
-  "figcaption",
-  "label",
-  "legend",
-  "fieldset",
-  "address",
-  "time",
-  "data",
-  "var",
-  "ruby",
-  "rt",
-  "rp",
-  "bdi",
-  "bdo",
-]);
-
-// HTML elements that manage DOM state (selection / focus / scroll /
-// playback / interactive identity). If any descendant of a keyed
-// element is one of these, reordering with an index key WOULD
-// corrupt that state. Detected via a bounded AST walk over the
-// JSXElement's children.
-const STATEFUL_HTML_DESCENDANT_TAGS: ReadonlySet<string> = new Set([
-  // Form controls — their value, focus, and selection are tied to
-  // DOM identity.
-  "input",
-  "textarea",
-  "select",
-  "option",
-  "optgroup",
-  "button",
-  "form",
-  "output",
-  "progress",
-  "meter",
-  // Media — playback state.
-  "video",
-  "audio",
-  "source",
-  "track",
-  // Document embeddings — independent state.
-  "iframe",
-  "embed",
-  "object",
-  // Interactive containers — `<a href>` carries focus/click identity.
-  "a",
-  "details",
-  "summary",
-  "dialog",
-  // Editable content.
-  "canvas",
-]);
-
-const isStatelessLeafJsxName = (jsxOpeningName: EsTreeNode): boolean => {
-  if (!isNodeOfType(jsxOpeningName, "JSXIdentifier")) return false;
-  return STATELESS_HTML_LEAF_TAGS.has(jsxOpeningName.name);
-};
-
-// Bounded walk over a JSXElement's children to detect any stateful
-// HTML descendant or any CUSTOM (PascalCase) component descendant.
-// Custom components are assumed potentially-stateful — we can't see
-// inside them.
-const STATEFUL_DESCENDANT_SCAN_BUDGET = 200;
-
-const containsStatefulDescendant = (jsxElement: EsTreeNode): boolean => {
-  let budget = STATEFUL_DESCENDANT_SCAN_BUDGET;
-  const stack: Array<EsTreeNode> = [jsxElement];
-  while (stack.length > 0) {
-    if (budget <= 0) return true; // Conservative bail-out — assume stateful.
-    budget -= 1;
-    const node = stack.pop()!;
-    if (isNodeOfType(node, "JSXElement")) {
-      const opening = (node as { openingElement: EsTreeNode }).openingElement;
-      const name = (opening as { name?: EsTreeNode }).name;
-      if (name && isNodeOfType(name, "JSXIdentifier")) {
-        const tagName = name.name;
-        const firstChar = tagName.charCodeAt(0);
-        const isUppercase = firstChar >= 65 && firstChar <= 90;
-        // PascalCase custom component — unknown state, treat as stateful.
-        if (isUppercase) return true;
-        if (STATEFUL_HTML_DESCENDANT_TAGS.has(tagName)) return true;
-      }
-      // Member-expression JSX (e.g. `<Foo.Bar />`) — custom; stateful.
-      if (name && isNodeOfType(name, "JSXMemberExpression")) return true;
-      // Walk children.
-      const children = (node as { children?: ReadonlyArray<EsTreeNode> }).children ?? [];
-      for (const child of children) stack.push(child);
-      continue;
-    }
-    if (isNodeOfType(node, "JSXFragment")) {
-      const children = (node as { children?: ReadonlyArray<EsTreeNode> }).children ?? [];
-      for (const child of children) stack.push(child);
-      continue;
-    }
-    if (isNodeOfType(node, "JSXExpressionContainer")) {
-      // `{anything}` — could render arbitrary JSX. Conservative: bail.
-      const expression = node.expression as EsTreeNode;
-      // Function-call children (`{renderX(...)}`, `{items.map(...)}`,
-      // `{children}`) are unknown — assume stateful.
-      if (
-        isNodeOfType(expression, "CallExpression") ||
-        isNodeOfType(expression, "Identifier") ||
-        isNodeOfType(expression, "MemberExpression")
-      ) {
-        return true;
-      }
-      // Conditional / logical / template — walk into.
-      stack.push(expression);
-      continue;
-    }
-    // Walk through ternary/logical expressions to find JSX branches.
-    if (isNodeOfType(node, "ConditionalExpression")) {
-      stack.push(node.consequent as EsTreeNode, node.alternate as EsTreeNode);
-      continue;
-    }
-    if (isNodeOfType(node, "LogicalExpression")) {
-      stack.push(node.left as EsTreeNode, node.right as EsTreeNode);
-      continue;
-    }
-    // JSXText, Literal — pure content, never stateful.
-  }
-  return false;
-};
+const isStatelessLeafJsxName = (jsxOpeningName: EsTreeNode): boolean =>
+  isNodeOfType(jsxOpeningName, "JSXIdentifier") &&
+  STATELESS_HTML_LEAF_TAGS.has(jsxOpeningName.name);
 
 // Recognises `<React.Fragment>` / `<Fragment>` / shorthand `<>` —
 // fragments carry no DOM identity and no internal state, so an index
