+++ packages/oxlint-plugin-react-doctor/src/plugin/rules/state-and-effects/no-adjust-state-on-prop-change.ts
@@ -63,4 +63,3 @@ export const noAdjustStateOnPropChange = defineRule<Rule>({
     },
   }),
 });
-
